Revert "Revert "Update some OJ answers""
This reverts commit d40a5ac9159e04d7c27b2bd468f91039cff1af86.

+++ 程序设计基础-LevOJ课程/P1125.c
@@ -1,10 +1,10 @@
 #include<stdio.h>
 int main(){
-	unsigned long long a[36]={0,1,2};
+	int a[36]={0,1,2};
 	int i=3,n;
 	for(;i<36;i++)
 		a[i]=a[i-1]+a[i-2];
 	while(~scanf("%d",&n))
-		printf("%lld\n",a[n]);
+		printf("%d\n",a[n]);
 	return 0;
 }
